Refactor allTrafficWidget feature tour into its own file.
